refactor(auth-plugin): extract OpenAPI helper to src/plugins/auth/openapi-enhance.ts (CP-4)
src/index.ts now imports OpenAPI directly from openapi-enhance.ts
(no re-export via auth-plugin.ts).
