+++ src/index.ts
@@ -19,7 +19,8 @@ import { organizationController } from "./modules/organizations";
 import { paymentsController } from "./modules/payments";
 import { registerPaymentListeners } from "./modules/payments/hooks/listeners";
 import { publicController } from "./modules/public";
-import { betterAuthPlugin, OpenAPI } from "./plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "./plugins/auth/auth-plugin";
+import { OpenAPI } from "./plugins/auth/openapi-enhance";
 import { cronPlugin } from "./plugins/cron/cron-plugin";
 import { errorPlugin } from "./plugins/errors/error-plugin";
 import { healthPlugin } from "./plugins/health/health-plugin";
